SectionListを別コンポーネントに移動
